backend/requirements.txt
asgiref==3.9.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
cachetools==5.5.2
certifi==2025.6.15
cffi==1.17.1
channels==4.3.1
channels_redis==4.3.0
charset-normalizer==3.4.2
constantly==23.10.4
cryptography==45.0.5
daphne==4.2.1
Django==5.2.4
django-cors-headers==4.7.0
django-oauth-toolkit==3.0.1
django-redis==6.0.0
djangorestframework==3.16.0
google-auth==2.40.3
gunicorn==23.0.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
jwcrypto==1.5.6
msgpack==1.1.1
mysqlclient==2.2.7
oauthlib==3.3.1
packaging==25.0
pillow==11.3.0
psutil==7.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pyOpenSSL==25.1.0
redis==6.4.0
requests==2.32.4
rsa==4.9.1
service-identity==24.2.0
sqlparse==0.5.3
Twisted==25.5.0
txaio==25.6.1
typing_extensions==4.14.1
urllib3==2.5.0
zope.interface==7.2
python-dotenv>=1.0.0,<2.0.0
